requirements.txt
asgiref==3.3.1
boto3==1.17.40
botocore==1.20.40
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.7
Django==3.1.7
django-cors-headers==3.7.0
django-dotenv==1.4.2
django-storages==1.11.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.6.0
drf-yasg==1.20.0
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.3
jmespath==0.10.0
MarkupSafe==1.1.1
packaging==20.9
paypal-checkout-serversdk==1.0.1
paypalhttp==1.0.0
Pillow==8.1.2
psycopg2==2.8.6
pycparser==2.20
PyJWT==2.0.1
pyOpenSSL==20.0.1
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
ruamel.yaml==0.17.0
ruamel.yaml.clib==0.2.2
s3transfer==0.3.6
SDK==1.0.0
simplejson==3.17.2
six==1.15.0
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.4
